[REMOVE] unwanted code

+++ app/src/main/java/com/light/finder/ui/lightfinder/DetailFragment.kt
@@ -131,11 +131,8 @@ class DetailFragment : BaseFragment() {
             val displayMetrics = it.resources.displayMetrics
             val dpHeight = displayMetrics.heightPixels
             viewPagerDetail.updateLayoutParams<ViewGroup.LayoutParams> {
-                height = (dpHeight / 2).toInt()
+                height = (dpHeight / 2)
             }
-//            val param: ViewGroup.MarginLayoutParams = livePreviewButton.layoutParams as ViewGroup.MarginLayoutParams
-//            param.bottomMargin = dpHeight / 12
-//            livePreviewButton.layoutParams = param
             bottomSheetBehavior.peekHeight = (dpHeight / 2.5).toInt()
         }
     }
